+++ controllers/SiteController.php
@@ -2,6 +2,7 @@
 
 namespace app\controllers;
 
+use app\modules\westnet\models\NotifyPayment;
 use Yii;
 use yii\filters\AccessControl;
 use app\components\web\Controller;
@@ -36,6 +37,11 @@ class SiteController extends Controller
         if(!Yii::$app->user->isGuest){
             if (Yii::$app->user->identity->hasRole('home_is_agenda', false)) {
                 return $this->redirect(['/agenda/default/index']);
+            }
+
+
+            if(Yii::$app->user->identity->hasRole('User-alert-new-no-verified-tranferences', false) && NotifyPayment::transferenceNotifyPaymentsNotVerifiedExists()) {
+                Yii::$app->session->addFlash('info', Yii::t('app', 'Theres one or more notify payments by transference not verified'));
             }
         }
         return $this->render('index');