+++ modules/westnet/views/notify-payment/index.php
@@ -3,6 +3,7 @@
 use yii\helpers\Html;
 use yii\grid\GridView;
 use yii\bootstrap\Collapse;
+use app\modules\checkout\models\PaymentMethod;
 
 /* @var $this yii\web\View */
 /* @var $searchModel app\modules\westnet\models\search\NotifyPaymentSearch */
@@ -10,6 +11,13 @@ use yii\bootstrap\Collapse;
 
 $this->title = Yii::t('app', 'Notify payments');
 $this->params['breadcrumbs'][] = $this->title;
+$payment_method_tranferencia = PaymentMethod::getTransferencia();
+if($payment_method_tranferencia) {
+    $payment_method_tranferencia_id = $payment_method_tranferencia->payment_method_id;
+} else {
+    $payment_method_tranferencia_id = 0;
+}
+
 ?>
 <div class="notify-payment-index">
 
@@ -58,7 +66,17 @@ $this->params['breadcrumbs'][] = $this->title;
             'from',
             [
                 'class' => 'app\components\grid\ActionColumn',
-                'template' => '{view}'
+                'template' => '{view}{verify}',
+                'buttons' => [
+                    'verify' => function($url, $model, $key) use ($payment_method_tranferencia_id) {
+                        if($payment_method_tranferencia_id == $model->payment_method_id && $model->verified == 0) {
+                            return Html::a(Yii::t('app', 'Mark as verified'), ['/westnet/notify-payment/verify', 'notify_payment_id' => $key], [
+                                'class' => 'btn btn-warning',
+                                'data-confirm' => Yii::t('app', 'Are you sure you want to mark it as verified?')
+                            ]);
+                        }
+                    }
+                ]
             ],
         ],
     ]); ?>